Refactor: Break down Navbar.tsx into smaller components
The Navbar.tsx file was refactored into smaller, more focused components to improve maintainability.
